requirements.txt
torch
torchvision
pillow
opencv-python
torchsummary
matplotlib